move recyclerview utils to subdirectory
