+++ app/src/main/java/jibreelpowell/com/softwords/utils/recyclerview/DataBoundViewHolder.kt
@@ -1,4 +1,4 @@
-package jibreelpowell.com.softwords.utils
+package jibreelpowell.com.softwords.utils.recyclerview
 
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
